source code
Create your own to-do list with PHP, including a front-end design and
back-end, allowing users to add and mark items as done. Use this PHP
to-do list as your own to-do list, or integrate with a PHP
authentication system to let users add their own to-do items to their
list.